Migrate Mutation.createToken
